+++ MultiPy/sender/templates/index.html
@@ -292,7 +292,7 @@
     document.getElementById('refreshSender')?.addEventListener('click', () => location.reload());
   </script>
   <script src="https://cdn.socket.io/4.7.4/socket.io.min.js"></script>
-    <script src="main.js"></script>
+  <script src="/static/main.js"></script>
 </body>
 
 </html>
